@@ -6,20 +6,11 @@
  */
 
 import './bootstrap';
+import CharacterPane from './components/CharacterPane.vue';
+import InstanceList from './components/InstanceList.vue';
+import SearchForm from './components/SearchForm.vue';
 import store from './store';
-
-window.Vue = require('vue');
-
-/**
- * The following block of code may be used to automatically register your
- * Vue components. It will recursively scan this directory for the Vue
- * components and automatically register them with their "basename".
- *
- * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
- */
-
-const files = require.context('./', true, /\.vue$/i);
-files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default));
+import Vue from 'vue';
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -27,7 +18,10 @@ files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-const app = new Vue({
+new Vue({
     el: '#app',
+    components: {
+        CharacterPane, InstanceList, SearchForm,
+    },
     store
 });
